chore: Removed sub file for tests and moved to async-local-context-manager.test.js
